added skybox support and cone hittable

+++ vec3.hpp
@@ -151,7 +151,7 @@ vec3 reflect(const vec3& v, const vec3& n) {
 vec3 refract(const vec3& uv, const vec3 n, float etai_over_etat) {
     auto cos_theta = fmin(dot(-uv, n), 1.0);
     vec3 r_out_perp = etai_over_etat * (uv + cos_theta*n);
-    vec3 r_out_parallel = -sqrt(fabs(1.0 - r_out_perp.length_squared())) * n;
+    vec3 r_out_parallel = -sqrtf(fabs(1.0 - r_out_perp.length_squared())) * n;
     return r_out_perp + r_out_parallel;
 }
